\name{isAvailable}
\alias{isAvailable}
\title{Check whether manipulate is available}
\description{
Check whether manipulate is available in the current front-end environment.
}
\usage{
isAvailable()
}
\value{
\code{TRUE} if manipulate is available, otherwise \code{FALSE}.
}
%- maybe also 'usage' for other objects documented here.
\details{
The manipulate package works only within the RStudio front-end.
}
